WIP: major refactoring
