+++ pyproject.toml
@@ -10,7 +10,7 @@ packages = [{include = "zarrnii"}]
 
 
 [tool.poetry.dependencies]
-python = ">=3.10,<4.0"
+python = ">=3.10,<3.12"
 zarr = "^2.17.0"
 nibabel = "^5.2.0"
 dask = "^2024.2.0"
@@ -18,7 +18,6 @@ numpy = "^1.26.4"
 scipy = "^1.12.0"
 pandas = "^2.2.0"
 ome-zarr = "^0.9.0"
-
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.2.0"
